requirements-dev.txt
# Зависимости для разработки

# Главные зависимости (то же что в llm_service/requirements.txt)
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1
requests==2.31.0
python-dotenv==1.0.0
openai==1.13.3
ollama==0.1.23
gunicorn==21.2.0

# Дополнительные для разработки
pytest==7.4.3
pytest-asyncio==0.23.2
black==23.12.1
flake8==6.1.0
mypy==1.7.1
pylint==3.0.3
autopep8==2.0.4
ipython==8.18.1
jupyter==1.0.0
